refactor: move store slices into their own files
